+++ src/components/organisms/ChallengeWriteCard/style.ts
@@ -1,0 +1,8 @@
+import Styled from 'styled-components';
+import { theme } from 'styled-tools';
+
+export const Header = Styled.div`
+    padding-top:100px;
+    text-align:center;
+    ${theme('font.dp3_eng')};
+`;
